27.06

+++ WF.EnrolleeApplication.DataAccess/WF.EnrolleeApplication.DataAccess/Services/EmployeePostService.cs
@@ -1,5 +1,7 @@
-﻿using System;
+﻿using NLog;
+using System;
 using System.Collections.Generic;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -17,6 +19,7 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// Объект для работы с данными
         /// </summary>
         private EnrolleeContext context;
+        private static Logger logger = LogManager.GetCurrentClassLogger();
         public EmployeePostService(string connectionString)
         {
             context = new EnrolleeContext(connectionString);
@@ -27,9 +30,30 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <param name="post">Объект сотрудника</param>
         public void DeleteEmployeePost(EmployeePost post)
         {
-            EmployeePost postToDelete = context.EmployeePost.FirstOrDefault(ep => ep.PostId == post.PostId);
-            context.EmployeePost.Remove(postToDelete);
-            context.SaveChanges();
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к удалению должности.");
+            try
+            {
+                logger.Debug($"Поиск записи должности для удаления. Удаляемый объект : {post.ToString()}.");
+                EmployeePost postToDelete = context.EmployeePost.FirstOrDefault(ep => ep.PostId == post.PostId);
+                if (postToDelete != null)
+                {
+                    context.EmployeePost.Remove(postToDelete);
+                    context.SaveChanges();
+                    logger.Debug("Удаление записи должности успешно завершено.");
+                }
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка удаления записи должности.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка удаления записи должности.");
+                logger.Error($"Ошибка — {ex.Message}.");
+            }
         }
         /// <summary>
         /// Получение должности по уникальному идентификатору
@@ -38,8 +62,28 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <returns>Возвращается объект должности сотрудника</returns>
         public EmployeePost GetEmployeePost(int id)
         {
-            EmployeePost postById = context.EmployeePost.AsNoTracking().FirstOrDefault(ep => ep.PostId == id);
-            return postById;
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к поиску должности.");
+            try
+            {
+                logger.Debug($"Поиск записи должности по уникальному идентификатору = {id}.");
+                EmployeePost postById = context.EmployeePost.AsNoTracking().FirstOrDefault(ep => ep.PostId == id);
+                if (postById != null) logger.Debug($"Поиск окончен. Искомая запись: {postById.ToString()}.");
+                return postById;
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка поиска записи должности.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка поиска записи должности.");
+                logger.Error($"Ошибка — {ex.Message}.");
+                return null;
+            }
         }
         /// <summary>
         /// Получение должности по наименованию должности
@@ -48,8 +92,28 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <returns>Возвращается объект должности сотрудника</returns>
         public EmployeePost GetEmployeePost(string name)
         {
-            EmployeePost postByName = context.EmployeePost.AsNoTracking().FirstOrDefault(ep => ep.Name == name);
-            return postByName;
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к поиску должности.");
+            try
+            {
+                logger.Debug($"Поиск записи должности по наименованию = {name}.");
+                EmployeePost postByName = context.EmployeePost.AsNoTracking().FirstOrDefault(ep => ep.Name == name);
+                if (postByName != null) logger.Debug($"Поиск окончен. Искомая запись: {postByName.ToString()}.");
+                return postByName;
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка поиска записи должности.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка поиска записи должности.");
+                logger.Error($"Ошибка — {ex.Message}.");
+                return null;
+            }
         }
         /// <summary>
         /// Получение списка должностей
@@ -57,8 +121,28 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <returns>Список должностей сотрудников</returns>
         public List<EmployeePost> GetEmployeePosts()
         {
-            List<EmployeePost> posts = context.EmployeePost.AsNoTracking().ToList();
-            return posts;
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к поиску списка должностей.");
+            try
+            {
+                logger.Debug($"Получение списка должностей.");
+                List<EmployeePost> posts = context.EmployeePost.AsNoTracking().ToList();
+                logger.Debug($"Поиск окончен. Количество записей: {posts.Count}.");
+                return posts;
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка получения списка должностей.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка получения списка должностей.");
+                logger.Error($"Ошибка — {ex.Message}.");
+                return null;
+            }
         }
         /// <summary>
         /// Добавление новой должности сотрудников
@@ -67,9 +151,29 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <returns>Возвращаем только что созданный объект</returns>
         public EmployeePost InsertEmployeePost(EmployeePost post)
         {
-            context.EmployeePost.Add(post);
-            context.SaveChanges();
-            return post;
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к добавлению должности");
+            try
+            {
+                logger.Debug($"Добавляемая запись: {post.ToString()}");
+                context.EmployeePost.Add(post);
+                context.SaveChanges();
+                logger.Debug($"Должность успешно добавлена.");
+                return post;
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка добавления должности.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка добавления должности.");
+                logger.Error($"Ошибка — {ex.Message}.");
+                return null;
+            }
         }
         /// <summary>
         /// Обновление должности
@@ -78,14 +182,34 @@ namespace WF.EnrolleeApplication.DataAccess.Services
         /// <returns>Возвращаем объект должности, с обновленными данными</returns>
         public EmployeePost UpdateEmployeePost(EmployeePost post)
         {
-            EmployeePost postToUpdate = context.EmployeePost.FirstOrDefault(ep => ep.PostId == post.PostId);
-            postToUpdate.Name = post.Name;
-            postToUpdate.Note = post.Note;
-            postToUpdate.RegistrationAllow = post.RegistrationAllow;
-            postToUpdate.EnrollAllow = post.EnrollAllow;
-            postToUpdate.DictionaryAllow = post.DictionaryAllow;
-            context.SaveChanges();
-            return postToUpdate;
+            logger.Trace("Попытка подключения к источнику данных.");
+            logger.Trace("Подготовка к обновлению должности.");
+            try
+            {
+                EmployeePost postToUpdate = context.EmployeePost.FirstOrDefault(ep => ep.PostId == post.PostId);
+                logger.Debug($"Текущая запись: {postToUpdate.ToString()}");
+                postToUpdate.Name = post.Name;
+                postToUpdate.Note = post.Note;
+                postToUpdate.RegistrationAllow = post.RegistrationAllow;
+                postToUpdate.EnrollAllow = post.EnrollAllow;
+                postToUpdate.DictionaryAllow = post.DictionaryAllow;
+                context.SaveChanges();
+                logger.Debug($"Новая запись: {postToUpdate.ToString()}");
+                return postToUpdate;
+            }
+            catch (SqlException sqlEx)
+            {
+                logger.Error("Ошибка редактирования должности.");
+                logger.Error($"Ошибка SQL Server — {sqlEx.Number}.");
+                logger.Error($"Сообщение об ошибке: {sqlEx.Message}.");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                logger.Error("Ошибка редактирования должности.");
+                logger.Error($"Ошибка — {ex.Message}.");
+                return null;
+            }
         }
     }
 }
